Remove react dependency from Podspec
With this dependency present, people cannot install this using Cocoapods. It attempts to install React 0.11.0. Adding a pod line for React in the podfile is not a good idea for people who started their projects with `react-native init` because you'll get double references to the same files. See the following issues:

https://github.com/invertase/react-native-firebase/issues/324